[fix]: refactor test code to common utils, raise rather than assert
